chore: Remove real time notifications used for testing
