+++ frontend/src/features/admin/AdminMenus/Overviews/KPICard/KPICard.jsx
@@ -1,0 +1,49 @@
+import { ArrowUpRight, ArrowDownRight} from 'lucide-react';
+
+const KPICard = ({ title, value, change, icon: Icon, loading }) => {
+  const isPositive = change >= 0;
+  
+  if (loading) {
+    return (
+      <div className="bg-white p-6 rounded-lg shadow-sm animate-pulse">
+        <div className="flex items-center justify-between">
+          <div className="flex-1">
+            <div className="h-4 bg-gray-200 rounded w-24 mb-2"></div>
+            <div className="h-8 bg-gray-200 rounded w-16"></div>
+          </div>
+          <div className="w-12 h-12 bg-gray-200 rounded-full"></div>
+        </div>
+        <div className="flex items-center mt-4">
+          <div className="h-4 bg-gray-200 rounded w-20"></div>
+        </div>
+      </div>
+    );
+  }
+
+  return (
+    <div className="bg-white p-6 rounded-lg shadow-sm hover:shadow-md transition-shadow">
+      <div className="flex items-center justify-between">
+        <div>
+          <p className="text-sm text-gray-500">{title}</p>
+          <h3 className="text-2xl font-bold mt-1">{value}</h3>
+        </div>
+        <div className={`p-3 rounded-full ${isPositive ? 'bg-green-100' : 'bg-red-100'}`}>
+          <Icon className={isPositive ? 'text-green-600' : 'text-red-600'} size={24} />
+        </div>
+      </div>
+      <div className="flex items-center mt-4">
+        {isPositive ? (
+          <ArrowUpRight size={20} className="text-green-500" />
+        ) : (
+          <ArrowDownRight size={20} className="text-red-500" />
+        )}
+        <span className={`ml-2 text-sm ${isPositive ? 'text-green-500' : 'text-red-500'}`}>
+          {Math.abs(change)}% {isPositive ? 'increase' : 'decrease'}
+        </span>
+        <span className="text-gray-500 text-sm ml-2">vs last month</span>
+      </div>
+    </div>
+  );
+};
+
+export default KPICard;
